app/requirements.txt
uvicorn==0.30.1
fastapi==0.111.0
python-dotenv==1.0.1
fastapi_keycloak==1.0.11
PyJWT==2.9.0
minio==7.2.10
cryptography==43.0.3
SQLAlchemy==2.0.36
pydantic==2.7 
sqlmodel==0.0.22 